Rename Guessing_Game_Code to Guessing_Game_Code1
This is a python code. If a is in one of the place values, it means the digit is not in the number, B means the digit is in the number, but currently is in the wrong spot, and C means that you have it on the correct spot AND the correct number! Remember, you only have 10 tries, so guess wisely!!! Good luck!